Update Dockerfile
New maintenance updates for Nextcloud Hub 8, 9 and 10 are available. As always, minor releases include stability and security improvements that are designed to be a safe and quick upgrade.
Update summary

On 17th April 2025, following 10th April u​pdates, we released out-of-band hotfixes for Nextcloud Hub 8, 9 and 10 to address the issue with installation of the new instances. While you could upgrade an existing Nextcloud instance to versions 29.0.15, 30.0.9 or 31.0.3, you could experience issues with installing them anew.